Let the user pick change
There can be multiple changes for the same commit. (Some instances of
Gerrit allows a commit to be pushed to multiple branches).

If multiple changes are found for the HEAD commit, let the user pick
which one to load.

Some extra arguments are passed to git ls-remote to do some filtering
server side.