Support remotes other than origin
When a git repo contains more than one remote, the user can pick the
remote to load code review for.

This also move all the logic of loading a review into the review
module. review.loadReview() starts a chain of Promises and callbacks
until any review is loaded. Any errors along the way are presented as
error messages to the user.